1. Add list<string> MetaBody::tags, which hold metadata in the key=value format; this is to workaround current serialization that errors out on map<string,string> in findType. 2. Add meanStress and vlumetricStrain to Triaxial for python access 3. add neverDamage flag to Brefcom which forces elastic behavior regardless off strain magnitude. 4. Fix some warnings in STLImporter regarding signedness 5. add __attribute__((unused)) to a variable in Wm3Matrix3 that was not used during optimized builds - to avoid warning.

+++ lib/import/STLImporter.hpp
@@ -49,7 +49,7 @@ class STLImporter {
 	int number_of_flat_edges() const { return nbFlatEdges; }
 
 	/// max number of facet vertices (by default 4)
-	int max_vertices_in_facet; 
+	size_t max_vertices_in_facet; 
 
 	void set_imported_stuff(bool v, bool e, bool f) {import_vertices=v; import_edges=e; import_facets=f;}
 	/// import vertices (by default true)
